+++ src/main/java/com/codenjoy/blog/controller/UIController.java
@@ -1,19 +1,15 @@
 package com.codenjoy.blog.controller;
 
-import com.codenjoy.blog.facade.PageFacade;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
 @RequestMapping
 @AllArgsConstructor
 public class UIController {
-
-    private final PageFacade pages;
 
     @GetMapping()
     public String main() {
@@ -28,19 +24,9 @@ public class UIController {
     }
 
     @GetMapping("/ui/page")
-    public String getPage(Model model, @RequestParam("fileName") String fileName) {
-        String content = pages.content(fileName);
-        addAttribute(model, "content", content);
-
+    public String getPage(Model model) {
         model.addAttribute("pageName", "content");
         model.addAttribute("reason", "markdown");
         return "layout";
-    }
-
-    private void addAttribute(Model model, String name, String content) {
-        model.addAttribute(name,
-                content == null
-                        ? "There's nothing here..."
-                        : content);
     }
 }